Moved `stage.rs` to `stages.rs` and changed to module
